refactor: List와 Grid의 상속 관계 수정 및 메서드 정리
- MarketCollectionViewCell의 삭제
- MarketGridCollectionViewCell과 MarketListCollectionViewCell의 상속을 제거하고 별도의 셀로 가져가게끔 수정
- configureImage 메서드를 분리하여 GridDataSource와 ListDataSource에서 각각 실행하도록 변경
