+++ OpenMarket/OpenMarket/View/MarketListCollectionViewCell.swift
@@ -7,14 +7,41 @@
 
 import UIKit
 
-final class MarketListCollectionViewCell: MarketCollectionViewCell {
+final class MarketListCollectionViewCell: UICollectionViewCell {
+    let imageView: UIImageView = {
+        let imageView = UIImageView()
+        return imageView
+    }()
+    
+    private let nameLabel: UILabel = {
+        let label = UILabel()
+        label.font = .boldSystemFont(ofSize: 20)
+        return label
+    }()
+    
+    private let priceLabel: UILabel = {
+        let label = UILabel()
+        label.font = .preferredFont(forTextStyle: .body)
+        label.textColor = .systemRed
+        label.adjustsFontSizeToFitWidth = true
+        label.numberOfLines = 2
+        return label
+    }()
+    
+    private let stockLabel: UILabel = {
+        let label = UILabel()
+        label.font = .preferredFont(forTextStyle: .body)
+        label.textColor = .systemGray
+        return label
+    }()
+    
     private let accessaryImageView: UIImageView = {
         let imageView = UIImageView()
         imageView.image = UIImage(systemName: "greaterthan")
         imageView.tintColor = .systemGray
         return imageView
     }()
-
+    
     private let horizontalStackView: UIStackView = {
         let stackView = UIStackView()
         stackView.axis = .horizontal
@@ -23,7 +50,7 @@ final class MarketListCollectionViewCell: MarketCollectionViewCell {
         stackView.translatesAutoresizingMaskIntoConstraints = false
         return stackView
     }()
-
+    
     private let subHorizontalStackView: UIStackView = {
         let stackView = UIStackView()
         stackView.axis = .horizontal
@@ -32,7 +59,7 @@ final class MarketListCollectionViewCell: MarketCollectionViewCell {
         stackView.translatesAutoresizingMaskIntoConstraints = false
         return stackView
     }()
-
+    
     private let verticalStackView: UIStackView = {
         let stackView = UIStackView()
         stackView.setContentHuggingPriority(.defaultLow, for: .horizontal)
@@ -42,7 +69,7 @@ final class MarketListCollectionViewCell: MarketCollectionViewCell {
         stackView.translatesAutoresizingMaskIntoConstraints = false
         return stackView
     }()
-
+    
     private let entireStackView: UIStackView = {
         let stackView = UIStackView()
         stackView.alignment = .fill
@@ -52,7 +79,34 @@ final class MarketListCollectionViewCell: MarketCollectionViewCell {
         stackView.translatesAutoresizingMaskIntoConstraints = false
         return stackView
     }()
-
+    
+    func configureCell(with item: Item, spacingType: String) {
+        self.nameLabel.text = item.productName
+        
+        if item.price == item.bargainPrice {
+            self.priceLabel.text = item.price
+            self.priceLabel.textColor = .systemGray
+        } else {
+            let price = item.price + spacingType + item.bargainPrice
+            let attributeString = NSMutableAttributedString(string: price)
+            
+            attributeString.addAttribute(.strikethroughStyle,
+                                         value: NSUnderlineStyle.single.rawValue,
+                                         range: NSMakeRange(0, item.price.count))
+            attributeString.addAttribute(.foregroundColor,
+                                         value: UIColor.systemGray,
+                                         range: NSMakeRange(item.price.count + 1, item.bargainPrice.count))
+            self.priceLabel.attributedText = attributeString
+        }
+        
+        if item.stock != "0" {
+            self.stockLabel.text = "잔여수량 : " + item.stock
+        } else {
+            self.stockLabel.text = "품절"
+            self.stockLabel.textColor = .systemOrange
+        }
+    }
+    
     private func arrangeSubView() {
         stockLabel.textAlignment = .right
         
@@ -91,10 +145,16 @@ final class MarketListCollectionViewCell: MarketCollectionViewCell {
     required init?(coder: NSCoder) {
         super.init(coder: coder)
     }
+    
+    override func prepareForReuse(){
+        super.prepareForReuse()
+        stockLabel.textColor = .systemGray
+        priceLabel.textColor = .systemRed
+    }
 }
 
 extension CALayer {
-    func addBottomBorder() {
+    fileprivate func addBottomBorder() {
         let border = CALayer()
         border.backgroundColor = UIColor.systemGray3.cgColor
