+++ OpenMarket/OpenMarket/ViewController/MarketCollectionViewController.swift
@@ -80,7 +80,19 @@ final class MarketCollectionViewController: UICollectionViewController {
     private func makeListDataSource() -> DataSource {
         let registration = UICollectionView.CellRegistration<MarketListCollectionViewCell, Item>.init { cell, indexPath, item in
             cell.configureCell(with: item,spacingType: " ")
-            self.configureImage(item, indexPath, cell)
+            self.sessionManager.receiveData(baseURL: item.productImage) { result in
+                switch result {
+                case .success(let data):
+                    guard let imageData = UIImage(data: data) else { return }
+                    
+                    DispatchQueue.main.async {
+                        guard indexPath == self.collectionView.indexPath(for: cell) else { return }
+                        cell.imageView.image = imageData
+                    }
+                case .failure(_):
+                    print("서버 통신 실패")
+                }
+            }
         }
         
         return DataSource(collectionView: collectionView) { collectionView, indexPath, item -> UICollectionViewCell? in
@@ -91,7 +103,19 @@ final class MarketCollectionViewController: UICollectionViewController {
     private func makeGridDataSource() -> DataSource {
         let registration = UICollectionView.CellRegistration<MarketGridCollectionViewCell, Item>.init { cell, indexPath, item in
             cell.configureCell(with: item,spacingType: "\n")
-            self.configureImage(item, indexPath, cell)
+            self.sessionManager.receiveData(baseURL: item.productImage) { result in
+                switch result {
+                case .success(let data):
+                    guard let imageData = UIImage(data: data) else { return }
+                    
+                    DispatchQueue.main.async {
+                        guard indexPath == self.collectionView.indexPath(for: cell) else { return }
+                        cell.imageView.image = imageData
+                    }
+                case .failure(_):
+                    print("서버 통신 실패")
+                }
+            }
         }
         
         return DataSource(collectionView: collectionView) { collectionView, indexPath, item -> UICollectionViewCell? in
@@ -99,21 +123,21 @@ final class MarketCollectionViewController: UICollectionViewController {
         }
     }
     
-    private func configureImage(_ item: Item, _ indexPath: IndexPath, _ cell: MarketCollectionViewCell) {
-        self.sessionManager.receiveData(baseURL: item.productImage) { result in
-            switch result {
-            case .success(let data):
-                guard let imageData = UIImage(data: data) else { return }
-                
-                DispatchQueue.main.async {
-                    guard indexPath == self.collectionView.indexPath(for: cell) else { return }
-                    cell.imageView.image = imageData
-                }
-            case .failure(_):
-                print("서버 통신 실패")
-            }
-        }
-    }
+    //   private func configureImage(_ item: Item, _ indexPath: IndexPath, _ cell: UICollectionViewCell) {
+    //        self.sessionManager.receiveData(baseURL: item.productImage) { result in
+    //            switch result {
+    //            case .success(let data):
+    //                guard let imageData = UIImage(data: data) else { return }
+    //
+    //                DispatchQueue.main.async {
+    //                    guard indexPath == self.collectionView.indexPath(for: cell) else { return }
+    //                    cell.imageView.image = imageData
+    //                }
+    //            case .failure(_):
+    //                print("서버 통신 실패")
+    //            }
+    //        }
+    //    }
     
     // MARK: Data & Snapshot
     private func applySnapshots() {
